test(components): move component tests into nested directories
