--- tests/unit/components/EmbeddableFooter.spec.ts
@@ -1,10 +1,0 @@
-import { shallowMount } from '@vue/test-utils'
-import { describe, expect, it } from 'vitest'
-import EmbeddableFooter from '@/components/EmbeddableFooter.vue'
-
-describe('EmbeddableFooter.vue', () => {
-  it('is a Vue instance', () => {
-    const wrapper = shallowMount(EmbeddableFooter)
-    expect(wrapper.vm).toBeTruthy()
-  })
-})
